fix(wt-perf): move timeline tests to lib + exclude bin from test compile
The `[lib] name = "wt_perf"` and `[[bin]] name = "wt-perf"` artifacts
both normalize to the same `wt_perf` basename. cargo-affected's runner
shim disambiguates same-basename test binaries via a marker probe, but
markers only fire for `kind = "test" | "bench" | "example"` — `lib` and
`bin` get none. Once #2558 added inline tests under `bin/wt-perf`, both
candidates appeared in cargo-affected's binary map without markers, so
any hash drift between list and run aborted the shim with
`basename fallback ambiguous`. The collect-affected job on main has
been broken since #2558 merged.

Two changes together fix it:

- Move `render_timeline` (and its `describe`/`duration_of` helpers) plus
  the inline tests from `src/main.rs` into `src/lib.rs`. The bin no
  longer carries any `#[cfg(test)]` items.
- Set `test = false` on the `[[bin]]` in Cargo.toml so cargo doesn't
  emit a test binary for it. This removes the bin from cargo-affected's
  binary map entirely, so even with hash drift the basename fallback
  finds a single candidate (the lib). The bin still gets compiled at
  `target/debug/wt-perf` because `tests/builds.rs` triggers it.
